refactor(iot): consolidate certificate utilities and simplify IoT service
- Move certificate generation and signing logic from local_ca to new x509_utils module for reusability across services
- Enables future reuse of x509_utils by ACM, API Gateway, and other services needing local certificate operations
- Consolidate iot_broker functionality into main iot service module
