+++ ministack/services/iot.py
@@ -26,23 +26,34 @@ issued client certificates remain valid across restarts.
 
 from __future__ import annotations
 
+import asyncio
+import base64
 import copy
 import hashlib
 import json
 import logging
 import os
 import re
+import struct
+import time
+import uuid
 from datetime import datetime, timezone
+from typing import Awaitable, Callable
 
-from ministack.core import local_ca
 from ministack.core.persistence import load_state
 from ministack.core.responses import (
     AccountScopedDict,
+    _request_account_id,
     error_response_json,
     get_account_id,
     get_region,
     json_response,
     new_uuid,
+)
+from ministack.core.x509_utils import (
+    generate_ca,
+    get_certificate_id,
+    sign_leaf_certificate,
 )
 
 logger = logging.getLogger("iot")
@@ -64,6 +75,34 @@ _thing_groups: AccountScopedDict = AccountScopedDict()
 _certificates: AccountScopedDict = AccountScopedDict()  # certificateId -> Certificate dict
 _policies: AccountScopedDict = AccountScopedDict()  # policyName -> Policy dict
 
+# Local CA state — lazily generated on first use, persisted across restarts.
+import threading
+
+_CA_LOCK = threading.Lock()
+_ca_cert_pem: str | None = None
+_ca_key_pem: str | None = None
+
+
+def _ensure_ca() -> tuple[str, str]:
+    """Return (cert_pem, key_pem), generating lazily on first use."""
+    global _ca_cert_pem, _ca_key_pem
+    if _ca_cert_pem is not None and _ca_key_pem is not None:
+        return _ca_cert_pem, _ca_key_pem
+    with _CA_LOCK:
+        if _ca_cert_pem is not None and _ca_key_pem is not None:
+            return _ca_cert_pem, _ca_key_pem
+        cert_pem, key_pem = generate_ca()
+        _ca_cert_pem = cert_pem
+        _ca_key_pem = key_pem
+        logger.info("Local CA: generated new self-signed root certificate")
+        return cert_pem, key_pem
+
+
+def get_ca_cert_pem() -> str:
+    """Return the CA certificate in PEM format. Generates the CA on first call."""
+    cert_pem, _ = _ensure_ca()
+    return cert_pem
+
 
 # ---------------------------------------------------------------------------
 # Persistence
@@ -77,11 +116,15 @@ def get_state() -> dict:
         "thing_groups": copy.deepcopy(_thing_groups),
         "certificates": copy.deepcopy(_certificates),
         "policies": copy.deepcopy(_policies),
-        "local_ca": local_ca.get_state(),
+        "ca": {"ca_cert_pem": _ca_cert_pem, "ca_key_pem": _ca_key_pem}
+        if _ca_cert_pem and _ca_key_pem
+        else {},
+        "mqtt_broker": _broker_get_state(),
     }
 
 
 def restore_state(data: dict | None) -> None:
+    global _ca_cert_pem, _ca_key_pem
     if not data:
         return
     _things.update(data.get("things", {}))
@@ -89,16 +132,28 @@ def restore_state(data: dict | None) -> None:
     _thing_groups.update(data.get("thing_groups", {}))
     _certificates.update(data.get("certificates", {}))
     _policies.update(data.get("policies", {}))
-    local_ca.restore_state(data.get("local_ca"))
+    ca_data = data.get("ca")
+    if ca_data:
+        cert = ca_data.get("ca_cert_pem")
+        key = ca_data.get("ca_key_pem")
+        if cert and key:
+            with _CA_LOCK:
+                _ca_cert_pem = cert
+                _ca_key_pem = key
+            logger.info("Local CA: restored from persisted state")
+    _broker_restore_state(data.get("mqtt_broker"))
 
 
 def reset() -> None:
+    global _ca_cert_pem, _ca_key_pem
     _things.clear()
     _thing_types.clear()
     _thing_groups.clear()
     _certificates.clear()
     _policies.clear()
-    local_ca.reset()
+    with _CA_LOCK:
+        _ca_cert_pem = None
+        _ca_key_pem = None
 
 
 try:
@@ -731,12 +786,15 @@ def _create_keys_and_certificate(qp: dict) -> tuple:
     """Generate a fresh keypair and sign a leaf certificate with the Local CA."""
     set_active = qp.get("setAsActive", "false").lower() == "true"
     try:
-        cert_pem, private_pem, public_pem = local_ca.sign_leaf_certificate(
-            common_name="AWS IoT Certificate"
+        ca_cert_pem, ca_key_pem = _ensure_ca()
+        cert_pem, private_pem, public_pem = sign_leaf_certificate(
+            ca_cert_pem=ca_cert_pem,
+            ca_key_pem=ca_key_pem,
+            common_name="AWS IoT Certificate",
         )
     except RuntimeError as e:
         return error_response_json("InternalFailureException", str(e), 503)
-    cert_id = local_ca.get_certificate_id(cert_pem)
+    cert_id = get_certificate_id(cert_pem)
     arn = _cert_arn(cert_id)
     record = {
         "certificateId": cert_id,
@@ -771,7 +829,7 @@ def _register_certificate(payload: dict, qp: dict) -> tuple:
     set_active = bool(payload.get("setAsActive", False))
     status = payload.get("status")
     try:
-        cert_id = local_ca.get_certificate_id(cert_pem)
+        cert_id = get_certificate_id(cert_pem)
     except Exception as e:
         return error_response_json(
             "CertificateValidationException",
@@ -791,7 +849,7 @@ def _register_certificate(payload: dict, qp: dict) -> tuple:
         "status": status or ("ACTIVE" if set_active else "INACTIVE"),
         "creationDate": _now_epoch(),
         "ownedBy": get_account_id(),
-        "caCertificateId": payload.get("caCertificatePem") and local_ca.get_certificate_id(payload["caCertificatePem"]) or None,
+        "caCertificateId": payload.get("caCertificatePem") and get_certificate_id(payload["caCertificatePem"]) or None,
         "attachedThings": [],
         "attachedPolicies": [],
     }
@@ -1216,10 +1274,746 @@ def _list_attached_policies(path: str, qp: dict) -> tuple:
 
 
 # ---------------------------------------------------------------------------
-# Helper exports for iot_data / iot_broker
+# Helper exports for iot_data
 # ---------------------------------------------------------------------------
 
 
 def lookup_certificate_by_id(cert_id: str) -> dict | None:
     """Return the Certificate record for a given certificateId in the current account, or None."""
     return _certificates.get(cert_id)
+
+
+# ===========================================================================
+# MQTT Broker — embedded MQTT 3.1.1 broker logic over WebSocket
+# ===========================================================================
+#
+# The broker owns a small in-process pub/sub registry plus an MQTT 3.1.1
+# framing layer used between the broker and WebSocket clients (per the AWS
+# WS-MQTT subprotocol).
+#
+# Architecture (mirrors Transfer Family's shared SFTP listener):
+#   Client → WebSocket (gateway port) → Bridge → in-memory pub/sub
+#
+# Multi-tenancy is enforced by transparent topic prefixing: every
+# PUBLISH/SUBSCRIBE topic seen on the wire is internally prefixed with the
+# caller's account_id before it hits the registry, and the prefix is
+# stripped on outbound delivery.
+
+_broker_logger = logging.getLogger("iot_broker")
+
+# ---------------------------------------------------------------------------
+# In-memory pub/sub registry
+# ---------------------------------------------------------------------------
+
+_subscriptions: dict[str, set["_Subscription"]] = {}
+_retained: dict[str, "_RetainedMessage"] = {}
+_connected_clients: dict[tuple[str, str], "_WSSession"] = {}
+_persistent_sessions: dict[tuple[str, str], "_PersistentSessionState"] = {}
+_broker_lock = asyncio.Lock()
+
+_SESSION_EXPIRY_SECONDS: int = int(os.environ.get("IOT_SESSION_EXPIRY_SECONDS", "3600"))
+_MAX_QUEUED_MESSAGES = 1000
+
+
+class _RetainedMessage:
+    __slots__ = ("payload", "qos", "topic", "ts")
+
+    def __init__(self, topic: str, payload: bytes, qos: int):
+        self.topic = topic
+        self.payload = payload
+        self.qos = qos
+        self.ts = time.time()
+
+
+# ---------------------------------------------------------------------------
+# Broker persistence (retained messages)
+# ---------------------------------------------------------------------------
+
+
+def _broker_get_state() -> dict:
+    retained_list = []
+    for topic, msg in _retained.items():
+        retained_list.append({
+            "topic": msg.topic,
+            "payload": base64.b64encode(msg.payload).decode("ascii"),
+            "qos": msg.qos,
+        })
+    return {"retained": retained_list}
+
+
+def _broker_restore_state(data: dict | None) -> None:
+    if not data:
+        return
+    for entry in data.get("retained", []):
+        topic = entry["topic"]
+        payload = base64.b64decode(entry["payload"])
+        qos = entry.get("qos", 0)
+        _retained[topic] = _RetainedMessage(topic, payload, qos)
+
+
+class _PersistentSessionState:
+    __slots__ = ("subscriptions", "queued_messages", "created_at")
+
+    def __init__(self, subscriptions: list[str], created_at: float):
+        self.subscriptions: list[str] = subscriptions
+        self.queued_messages: list[tuple[str, bytes, int]] = []
+        self.created_at: float = created_at
+
+
+def _is_session_expired(session_state: _PersistentSessionState) -> bool:
+    return (time.time() - session_state.created_at) > _SESSION_EXPIRY_SECONDS
+
+
+class _InFlightMessage:
+    __slots__ = ("packet_id", "topic", "payload", "sent_at", "retransmit_count")
+
+    def __init__(self, packet_id: int, topic: str, payload: bytes):
+        self.packet_id = packet_id
+        self.topic = topic
+        self.payload = payload
+        self.sent_at = asyncio.get_event_loop().time()
+        self.retransmit_count = 0
+
+
+_RETRANSMIT_INTERVAL_SECONDS = int(os.environ.get("IOT_RETRANSMIT_SECONDS", "10"))
+
+
+class _Subscription:
+    __slots__ = ("subscription_id", "filter_prefixed", "account_id", "deliver", "granted_qos")
+
+    def __init__(
+        self,
+        filter_prefixed: str,
+        account_id: str,
+        deliver: Callable[[str, bytes, int], Awaitable[None]],
+        granted_qos: int = 0,
+    ):
+        self.subscription_id = uuid.uuid4().hex
+        self.filter_prefixed = filter_prefixed
+        self.account_id = account_id
+        self.deliver = deliver
+        self.granted_qos = granted_qos
+
+    def __hash__(self) -> int:
+        return hash(self.subscription_id)
+
+    def __eq__(self, other) -> bool:
+        return isinstance(other, _Subscription) and other.subscription_id == self.subscription_id
+
+
+# ---------------------------------------------------------------------------
+# Topic prefixing & matching
+# ---------------------------------------------------------------------------
+
+
+def _scoped_topic(account_id: str, topic: str) -> str:
+    return f"{account_id}/{topic}"
+
+
+def _unscope_topic(account_id: str, scoped_topic: str) -> str:
+    prefix = f"{account_id}/"
+    if scoped_topic.startswith(prefix):
+        return scoped_topic[len(prefix):]
+    return scoped_topic
+
+
+def _topic_matches(filter_: str, topic: str) -> bool:
+    f_parts = filter_.split("/")
+    t_parts = topic.split("/")
+    fi = ti = 0
+    while fi < len(f_parts):
+        f = f_parts[fi]
+        if f == "#":
+            return True
+        if ti >= len(t_parts):
+            return False
+        if f != "+" and f != t_parts[ti]:
+            return False
+        fi += 1
+        ti += 1
+    return ti == len(t_parts)
+
+
+# ---------------------------------------------------------------------------
+# Topic validation
+# ---------------------------------------------------------------------------
+
+_MQTT_MAX_TOPIC_BYTES = 256
+
+
+def _validate_publish_topic(topic: str) -> bool:
+    if not topic:
+        return False
+    if "+" in topic or "#" in topic:
+        return False
+    if len(topic.encode("utf-8")) > _MQTT_MAX_TOPIC_BYTES:
+        return False
+    return True
+
+
+# ---------------------------------------------------------------------------
+# Broker public API (consumed by iot_data.py and handle_websocket)
+# ---------------------------------------------------------------------------
+
+
+def broker_is_available() -> bool:
+    return True
+
+
+async def broker_start() -> None:
+    return None
+
+
+async def broker_stop() -> None:
+    async with _broker_lock:
+        _subscriptions.clear()
+        _retained.clear()
+        _connected_clients.clear()
+        _persistent_sessions.clear()
+
+
+async def broker_publish(
+    account_id: str,
+    topic: str,
+    payload: bytes,
+    qos: int = 0,
+    retain: bool = False,
+) -> None:
+    scoped = _scoped_topic(account_id, topic)
+
+    if retain:
+        if not payload:
+            _retained.pop(scoped, None)
+        else:
+            _retained[scoped] = _RetainedMessage(scoped, payload, qos)
+
+    async with _broker_lock:
+        subs = [s for sset in _subscriptions.values() for s in sset]
+
+    for sub in subs:
+        if _topic_matches(sub.filter_prefixed, scoped):
+            try:
+                effective_qos = min(qos, sub.granted_qos)
+                await sub.deliver(_unscope_topic(sub.account_id, scoped), payload, effective_qos)
+            except Exception:
+                _broker_logger.exception("IoT broker: subscriber %s delivery failed", sub.subscription_id)
+
+    if qos >= 1:
+        for key, ps in list(_persistent_sessions.items()):
+            ps_account_id, ps_client_id = key
+            if ps_account_id != account_id:
+                continue
+            if key in _connected_clients:
+                continue
+            if _is_session_expired(ps):
+                continue
+            for filt in ps.subscriptions:
+                scoped_filter = _scoped_topic(ps_account_id, filt)
+                if _topic_matches(scoped_filter, scoped):
+                    ps.queued_messages.append((topic, payload, qos))
+                    if len(ps.queued_messages) > _MAX_QUEUED_MESSAGES:
+                        ps.queued_messages = ps.queued_messages[-_MAX_QUEUED_MESSAGES:]
+                    break
+
+
+async def broker_subscribe(
+    account_id: str,
+    topic_filter: str,
+    callback: Callable[[str, bytes, int], Awaitable[None]],
+    granted_qos: int = 0,
+) -> str:
+    filter_prefixed = _scoped_topic(account_id, topic_filter)
+    sub = _Subscription(filter_prefixed, account_id, callback, granted_qos)
+    async with _broker_lock:
+        _subscriptions.setdefault(filter_prefixed, set()).add(sub)
+        has_wildcard = "+" in topic_filter or "#" in topic_filter
+        if not has_wildcard:
+            retained_to_send = [
+                r for k, r in _retained.items() if _topic_matches(filter_prefixed, k)
+            ]
+        else:
+            retained_to_send = []
+
+    for r in retained_to_send:
+        try:
+            await sub.deliver(_unscope_topic(account_id, r.topic), r.payload, r.qos)
+        except Exception:
+            _broker_logger.exception("IoT broker: retained-message delivery failed")
+
+    return sub.subscription_id
+
+
+async def broker_unsubscribe(subscription_id: str) -> None:
+    async with _broker_lock:
+        for filter_, subs in list(_subscriptions.items()):
+            for s in list(subs):
+                if s.subscription_id == subscription_id:
+                    subs.discard(s)
+            if not subs:
+                _subscriptions.pop(filter_, None)
+
+
+def broker_reset() -> None:
+    _subscriptions.clear()
+    _retained.clear()
+    _connected_clients.clear()
+    _persistent_sessions.clear()
+
+
+# ---------------------------------------------------------------------------
+# Connected-client registry & duplicate detection
+# ---------------------------------------------------------------------------
+
+
+def _register_client(account_id: str, client_id: str, session: "_WSSession") -> None:
+    _connected_clients[(account_id, client_id)] = session
+
+
+def _deregister_client(account_id: str, client_id: str) -> None:
+    _connected_clients.pop((account_id, client_id), None)
+
+
+async def _force_disconnect_duplicate(account_id: str, client_id: str) -> None:
+    key = (account_id, client_id)
+    existing = _connected_clients.get(key)
+    if existing is not None:
+        _broker_logger.info("IoT broker: duplicate client_id=%s, forcing old connection closed", client_id)
+        try:
+            await existing._send({"type": "websocket.close", "code": 1000})
+        except Exception:
+            pass
+        await existing.cleanup()
+        _connected_clients.pop(key, None)
+
+
+# ---------------------------------------------------------------------------
+# MQTT 3.1.1 frame codec
+# ---------------------------------------------------------------------------
+
+PKT_CONNECT = 1
+PKT_CONNACK = 2
+PKT_PUBLISH = 3
+PKT_PUBACK = 4
+PKT_SUBSCRIBE = 8
+PKT_SUBACK = 9
+PKT_UNSUBSCRIBE = 10
+PKT_UNSUBACK = 11
+PKT_PINGREQ = 12
+PKT_PINGRESP = 13
+PKT_DISCONNECT = 14
+
+
+def _encode_remaining_length(n: int) -> bytes:
+    out = bytearray()
+    while True:
+        byte = n & 0x7F
+        n >>= 7
+        if n > 0:
+            byte |= 0x80
+        out.append(byte)
+        if n == 0:
+            return bytes(out)
+
+
+def _decode_remaining_length(buf: bytes, offset: int) -> tuple[int, int]:
+    multiplier = 1
+    value = 0
+    pos = offset
+    while True:
+        if pos >= len(buf):
+            raise ValueError("Truncated remaining length")
+        b = buf[pos]
+        pos += 1
+        value += (b & 0x7F) * multiplier
+        if b & 0x80 == 0:
+            break
+        multiplier *= 128
+        if multiplier > 128 * 128 * 128:
+            raise ValueError("Remaining length exceeds 4 bytes")
+    return value, pos
+
+
+def _read_string(buf: bytes, offset: int) -> tuple[str, int]:
+    if offset + 2 > len(buf):
+        raise ValueError("Truncated string length")
+    n = struct.unpack_from("!H", buf, offset)[0]
+    offset += 2
+    if offset + n > len(buf):
+        raise ValueError("Truncated string body")
+    return buf[offset:offset + n].decode("utf-8"), offset + n
+
+
+def _encode_string(s: str) -> bytes:
+    raw = s.encode("utf-8")
+    return struct.pack("!H", len(raw)) + raw
+
+
+def _make_connack(return_code: int = 0, session_present: bool = False) -> bytes:
+    flags = 1 if session_present else 0
+    body = bytes([flags, return_code])
+    return bytes([PKT_CONNACK << 4]) + _encode_remaining_length(len(body)) + body
+
+
+def _make_publish(topic: str, payload: bytes, qos: int = 0, packet_id: int | None = None,
+                  retain: bool = False, dup: bool = False) -> bytes:
+    fixed = (PKT_PUBLISH << 4) | (qos << 1) | (0x08 if dup else 0) | (0x01 if retain else 0)
+    body = _encode_string(topic)
+    if qos > 0:
+        if packet_id is None:
+            packet_id = 1
+        body += struct.pack("!H", packet_id)
+    body += payload
+    return bytes([fixed]) + _encode_remaining_length(len(body)) + body
+
+
+def _make_puback(packet_id: int) -> bytes:
+    return bytes([PKT_PUBACK << 4]) + bytes([2]) + struct.pack("!H", packet_id)
+
+
+def _make_suback(packet_id: int, granted_qos: list[int]) -> bytes:
+    body = struct.pack("!H", packet_id) + bytes(granted_qos)
+    return bytes([PKT_SUBACK << 4]) + _encode_remaining_length(len(body)) + body
+
+
+def _make_unsuback(packet_id: int) -> bytes:
+    return bytes([PKT_UNSUBACK << 4]) + bytes([2]) + struct.pack("!H", packet_id)
+
+
+def _make_pingresp() -> bytes:
+    return bytes([PKT_PINGRESP << 4, 0])
+
+
+# ---------------------------------------------------------------------------
+# WebSocket session driver
+# ---------------------------------------------------------------------------
+
+
+def _max_frame_buffer_bytes() -> int:
+    return int(os.environ.get("IOT_WS_FRAME_MAX_BYTES", str(16 * 1024 * 1024)))
+
+
+class _WSSession:
+    def __init__(self, send_coro, account_id: str):
+        self._send = send_coro
+        self.account_id = account_id
+        self._sub_ids: list[str] = []
+        self._sub_filters: dict[str, str] = {}
+        self._sub_granted_qos: dict[str, int] = {}
+        self._buffer = bytearray()
+        self._next_pid = 1
+        self._send_lock = asyncio.Lock()
+        self._client_id: str = ""
+        self._clean_session: bool = True
+        self._in_flight: dict[int, _InFlightMessage] = {}
+        self._retransmit_task: asyncio.Task | None = None
+        self._will_topic: str | None = None
+        self._will_message: bytes | None = None
+        self._will_qos: int = 0
+        self._will_retain: bool = False
+        self._graceful_disconnect: bool = False
+
+    def _alloc_packet_id(self) -> int:
+        pid = self._next_pid
+        self._next_pid = (self._next_pid % 65535) + 1
+        return pid
+
+    def _ensure_retransmit_timer(self) -> None:
+        if self._retransmit_task is None or self._retransmit_task.done():
+            self._retransmit_task = asyncio.ensure_future(self._retransmit_loop())
+
+    async def _retransmit_loop(self) -> None:
+        try:
+            while self._in_flight:
+                await asyncio.sleep(_RETRANSMIT_INTERVAL_SECONDS)
+                now = asyncio.get_event_loop().time()
+                for pid, msg in list(self._in_flight.items()):
+                    if now - msg.sent_at >= _RETRANSMIT_INTERVAL_SECONDS:
+                        msg.retransmit_count += 1
+                        msg.sent_at = now
+                        await self.send_bytes(
+                            _make_publish(msg.topic, msg.payload, qos=1, packet_id=pid, dup=True)
+                        )
+        except asyncio.CancelledError:
+            pass
+
+    async def send_bytes(self, b: bytes) -> None:
+        async with self._send_lock:
+            await self._send({"type": "websocket.send", "bytes": b})
+
+    async def deliver_to_client(self, topic: str, payload: bytes, qos: int) -> None:
+        if qos == 0:
+            await self.send_bytes(_make_publish(topic, payload, qos=0))
+        else:
+            pid = self._alloc_packet_id()
+            self._in_flight[pid] = _InFlightMessage(pid, topic, payload)
+            await self.send_bytes(_make_publish(topic, payload, qos=1, packet_id=pid))
+            self._ensure_retransmit_timer()
+
+    def _take_packet(self) -> tuple[int, int, bytes] | None:
+        if len(self._buffer) < 2:
+            return None
+        first = self._buffer[0]
+        try:
+            remaining, header_end = _decode_remaining_length(bytes(self._buffer), 1)
+        except ValueError:
+            if len(self._buffer) > 5:
+                self._buffer.clear()
+            return None
+        total = header_end + remaining
+        if len(self._buffer) < total:
+            return None
+        body = bytes(self._buffer[header_end:total])
+        del self._buffer[:total]
+        pkt_type = (first >> 4) & 0x0F
+        flags = first & 0x0F
+        return pkt_type, flags, body
+
+    async def handle_packet(self, pkt_type: int, flags: int, body: bytes) -> bool:
+        if pkt_type == PKT_CONNECT:
+            off = 0
+            _proto_name, off = _read_string(body, off)
+            off += 1  # Protocol Level
+            if off >= len(body):
+                await self.send_bytes(_make_connack(return_code=0))
+                return True
+            connect_flags = body[off]
+            off += 1
+            off += 2  # Keep Alive
+
+            will_flag = bool(connect_flags & 0x04)
+            will_qos = (connect_flags >> 3) & 0x03
+            will_retain = bool(connect_flags & 0x20)
+            clean_session = bool(connect_flags & 0x02)
+
+            self._clean_session = clean_session
+
+            if off < len(body):
+                client_id, off = _read_string(body, off)
+            else:
+                client_id = ""
+            if not client_id:
+                client_id = uuid.uuid4().hex
+            self._client_id = client_id
+
+            if will_flag:
+                if off < len(body):
+                    will_topic, off = _read_string(body, off)
+                else:
+                    will_topic = ""
+                if off + 2 <= len(body):
+                    msg_len = struct.unpack_from("!H", body, off)[0]
+                    off += 2
+                    will_message = body[off:off + msg_len]
+                    off += msg_len
+                else:
+                    will_message = b""
+                self._will_topic = will_topic
+                self._will_message = will_message
+                self._will_qos = will_qos
+                self._will_retain = will_retain
+            else:
+                self._will_topic = None
+                self._will_message = None
+                self._will_qos = 0
+                self._will_retain = False
+
+            self._graceful_disconnect = False
+            await _force_disconnect_duplicate(self.account_id, self._client_id)
+            _register_client(self.account_id, self._client_id, self)
+
+            session_key = (self.account_id, self._client_id)
+            session_present = False
+
+            if clean_session:
+                _persistent_sessions.pop(session_key, None)
+            else:
+                existing_ps = _persistent_sessions.get(session_key)
+                if existing_ps is not None and not _is_session_expired(existing_ps):
+                    session_present = True
+                    for topic_filter in existing_ps.subscriptions:
+                        sid = await broker_subscribe(
+                            self.account_id, topic_filter, self.deliver_to_client, 1
+                        )
+                        self._sub_ids.append(sid)
+                        self._sub_filters[sid] = topic_filter
+                        self._sub_granted_qos[sid] = 1
+                    await self.send_bytes(_make_connack(return_code=0, session_present=True))
+                    queued = existing_ps.queued_messages[:]
+                    existing_ps.queued_messages.clear()
+                    for q_topic, q_payload, q_qos in queued:
+                        await self.deliver_to_client(q_topic, q_payload, q_qos)
+                    return True
+                else:
+                    _persistent_sessions[session_key] = _PersistentSessionState(
+                        subscriptions=[], created_at=time.time()
+                    )
+
+            await self.send_bytes(_make_connack(return_code=0, session_present=session_present))
+            return True
+
+        if pkt_type == PKT_PUBLISH:
+            qos = (flags >> 1) & 0x03
+            retain = bool(flags & 0x01)
+            topic, off = _read_string(body, 0)
+            packet_id = None
+            if qos > 0:
+                if off + 2 > len(body):
+                    return True
+                packet_id = struct.unpack_from("!H", body, off)[0]
+                off += 2
+            if not _validate_publish_topic(topic):
+                _broker_logger.warning("IoT broker: PUBLISH rejected — invalid topic: %r", topic)
+                return False
+            payload = body[off:]
+            await broker_publish(self.account_id, topic, payload, qos=qos, retain=retain)
+            if qos == 1 and packet_id is not None:
+                await self.send_bytes(_make_puback(packet_id))
+            return True
+
+        if pkt_type == PKT_SUBSCRIBE:
+            packet_id = struct.unpack_from("!H", body, 0)[0]
+            off = 2
+            granted = []
+            while off < len(body):
+                topic, off = _read_string(body, off)
+                req_qos = body[off]
+                off += 1
+                granted_qos = min(req_qos, 1)
+                granted.append(granted_qos)
+                sid = await broker_subscribe(self.account_id, topic, self.deliver_to_client, granted_qos)
+                self._sub_ids.append(sid)
+                self._sub_filters[sid] = topic
+                self._sub_granted_qos[sid] = granted_qos
+            await self.send_bytes(_make_suback(packet_id, granted))
+            return True
+
+        if pkt_type == PKT_PUBACK:
+            if len(body) >= 2:
+                packet_id = struct.unpack_from("!H", body, 0)[0]
+                self._in_flight.pop(packet_id, None)
+            return True
+
+        if pkt_type == PKT_UNSUBSCRIBE:
+            packet_id = struct.unpack_from("!H", body, 0)[0]
+            for sid in list(self._sub_ids):
+                await broker_unsubscribe(sid)
+            self._sub_ids.clear()
+            await self.send_bytes(_make_unsuback(packet_id))
+            return True
+
+        if pkt_type == PKT_PINGREQ:
+            await self.send_bytes(_make_pingresp())
+            return True
+
+        if pkt_type == PKT_DISCONNECT:
+            self._graceful_disconnect = True
+            return False
+
+        return True
+
+    async def cleanup(self) -> None:
+        if self._retransmit_task is not None and not self._retransmit_task.done():
+            self._retransmit_task.cancel()
+            try:
+                await self._retransmit_task
+            except asyncio.CancelledError:
+                pass
+            self._retransmit_task = None
+        self._in_flight.clear()
+        if not self._graceful_disconnect and self._will_topic is not None:
+            await broker_publish(
+                self.account_id,
+                self._will_topic,
+                self._will_message or b"",
+                qos=self._will_qos,
+                retain=self._will_retain,
+            )
+        if not self._clean_session and self._client_id:
+            self._preserve_session()
+        for sid in self._sub_ids:
+            await broker_unsubscribe(sid)
+        self._sub_ids.clear()
+        self._sub_filters.clear()
+        self._sub_granted_qos.clear()
+        if self._client_id:
+            _deregister_client(self.account_id, self._client_id)
+
+    def _preserve_session(self) -> None:
+        session_key = (self.account_id, self._client_id)
+        unprefixed_filters = list(self._sub_filters.values())
+        existing = _persistent_sessions.get(session_key)
+        if existing is not None:
+            existing.subscriptions = unprefixed_filters
+            existing.created_at = time.time()
+        else:
+            _persistent_sessions[session_key] = _PersistentSessionState(
+                subscriptions=unprefixed_filters, created_at=time.time()
+            )
+
+
+async def handle_websocket(scope: dict, receive, send, account_id: str) -> None:
+    """Drive an MQTT-over-WebSocket session."""
+    msg = await receive()
+    if msg.get("type") != "websocket.connect":
+        return
+
+    sub_headers = {}
+    for name, value in scope.get("headers", []):
+        try:
+            sub_headers[name.decode("latin-1").lower()] = value.decode("utf-8")
+        except UnicodeDecodeError:
+            sub_headers[name.decode("latin-1").lower()] = value.decode("latin-1")
+    requested = sub_headers.get("sec-websocket-protocol", "")
+    chosen = None
+    for proto in [p.strip() for p in requested.split(",") if p.strip()]:
+        if proto.lower() in ("mqtt", "mqttv3.1", "mqttv5"):
+            chosen = proto
+            break
+
+    accept: dict = {"type": "websocket.accept"}
+    if chosen:
+        accept["subprotocol"] = chosen
+    await send(accept)
+
+    ctx_token = _request_account_id.set(account_id)
+    session = _WSSession(send, account_id)
+    max_buffer = _max_frame_buffer_bytes()
+
+    try:
+        while True:
+            incoming = await receive()
+            mtype = incoming.get("type")
+            if mtype == "websocket.disconnect":
+                break
+            if mtype != "websocket.receive":
+                continue
+            data = incoming.get("bytes")
+            if data is None:
+                text = incoming.get("text")
+                if text is None:
+                    continue
+                continue
+            session._buffer.extend(data)
+            if len(session._buffer) > max_buffer:
+                _broker_logger.warning("IoT broker: WS buffer overflow, dropping connection")
+                break
+            while True:
+                pkt = session._take_packet()
+                if pkt is None:
+                    break
+                pkt_type, flags, body = pkt
+                cont = await session.handle_packet(pkt_type, flags, body)
+                if not cont:
+                    return
+    except Exception:
+        _broker_logger.exception("IoT broker WebSocket session failed")
+    finally:
+        await session.cleanup()
+        try:
+            _request_account_id.reset(ctx_token)
+        except Exception:
+            pass
+        try:
+            await send({"type": "websocket.close", "code": 1000})
+        except Exception:
+            pass
