+++ tests/test_iot_lwt.py
@@ -3,13 +3,13 @@
 import asyncio
 import struct
 
-from ministack.services.iot_broker import (
+from ministack.services.iot import (
     PKT_CONNECT,
     PKT_DISCONNECT,
     _WSSession,
     _encode_string,
-    reset,
-    subscribe,
+    broker_reset as reset,
+    broker_subscribe as subscribe,
 )
